requirements.txt
nltk==3.5
pandas==1.0.1
numpy==1.16.3
praw==7.0.0
matplotlib==3.0.2
scipy==1.2.1
regex==2020.5.7
progressbar33==2.4
